Code review: Code moved.
